In this project, we will delve into the data of hotel bookings. The dataset at hand provides details about bookings in both city and resort hotels. Information like booking dates, duration of stays, counts of adults, children, babies, and available parking spaces are included.
The hospitality industry is notably dynamic, with bookings influenced by multiple variables, among others. Our focus is to uncover patterns and insights hidden within the data.

The primary aim of this endeavor is to thoroughly analyze the dataset, aiming to unveil pivotal factors that influence hotel bookings. By doing so, we aspire to offer valuable insights to hotel management. These insights could pave the way for strategic campaigns that enhance business prospects and overall performance. Through the lens of data exploration, we endeavor to decipher the complex landscape of hotel bookings, empowering decision-makers to make informed choices and ultimately elevate the hotel's success trajectory.